Move the exporter in the kinto2yaml module.
