+++ kinto_wizard/logger.py
@@ -1,0 +1,2 @@
+import logging
+logger = logging.getLogger("kinto-wizard")
