+++ tests/kinto.ini
@@ -46,7 +46,7 @@ kinto.flush_endpoint_enabled = true
 # Plugins
 #
 kinto.includes = kinto.plugins.default_bucket
-
+#                 kinto.plugins.admin
 #
 # Firefox Accounts configuration.
 #   These are working FxA credentials for localhost:8888
